% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{massqc_conflicts}
\alias{massqc_conflicts}
\title{Conflicts between the massqc and other packages}
\usage{
massqc_conflicts()
}
\value{
massqc_conflicts
}
\description{
This function lists all the conflicts between packages in the massqc
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} 
from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
massqc_conflicts()
}
