% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massqc_rsd_plot.R
\name{massqc_cumulative_rsd_plot}
\alias{massqc_cumulative_rsd_plot}
\title{massqc_rsd_plot}
\usage{
massqc_cumulative_rsd_plot(..., title, rsd_cutoff, color)
}
\arguments{
\item{...}{one or more mass_dataset object}

\item{title}{missing or same length (character vector) with ...}

\item{rsd_cutoff}{numeric}

\item{color}{when one mass_dataset, the line color}
}
\value{
ggplot2 plot.
}
\description{
RSD plot for each variable.
}
\examples{
library(massdataset)
library(ggplot2)
library(tidyverse)
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )

massqc_cumulative_rsd_plot(object, rsd_cutoff = 30, color = "blue")

object1 =
  object \%>\%
  activate_mass_dataset(what = "sample_info") \%>\%
  dplyr::filter(class == "QC")

object2 =
  object \%>\%
  activate_mass_dataset(what = "sample_info") \%>\%
  dplyr::filter(class == "Subject")

massqc_cumulative_rsd_plot(object1, object2,
                           title = c("QC", "Subject")) +
  ggsci::scale_color_lancet()

massqc_cumulative_rsd_plot(object1,
                           object2,
                           rsd_cutoff = 30,
                           title = c("QC", "Subject")) +
  ggsci::scale_color_lancet()

massqc_cumulative_rsd_plot(
  object1,
  title = c("QC"),
  color = "red",
  rsd_cutoff = 30
)
massqc_cumulative_rsd_plot(
  object2,
  title = c("Subject"),
  color = "blue",
  rsd_cutoff = 50
)
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
