% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{massqc_packages}
\alias{massqc_packages}
\title{List all packages in the massqc}
\usage{
massqc_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include massqc in the list?}
}
\description{
List all packages in the massqc
}
