% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massqc_pca.R
\name{massqc_pca}
\alias{massqc_pca}
\title{massqc_pca}
\usage{
massqc_pca(
  object,
  color_by,
  point_alpha = 0.8,
  frame = TRUE,
  frame.type = "norm",
  ...
)
}
\arguments{
\item{object}{tidymass-class object.}

\item{color_by}{which column (sample_info) is used to color samples}

\item{point_alpha}{point_alpha}

\item{frame}{?ggplot2::autoplot}

\item{frame.type}{?ggplot2::autoplot}

\item{...}{other paramters for ggplot2::autoplot}
}
\value{
ggplot2 plot.
}
\description{
Boxplot for each samples.
}
\examples{
library(massdataset)
library(ggplot2)
library(tidyverse)
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )

object \%>\%
  massqc_pca()

object \%>\%
  massqc_pca(color_by = "class")

object \%>\%
  scale \%>\%
  massqc_pca(color_by = "class")

object \%>\%
  scale \%>\%
  massqc_pca(color_by = "class", frame = FALSE) +
  ggsci::scale_fill_lancet()

object \%>\%
  scale \%>\%
  massqc_pca(color_by = "class", frame = FALSE) +
  ggsci::scale_fill_lancet() +
  ggrepel::geom_text_repel(aes(label = sample_id))

object \%>\%
  scale \%>\%
  massqc_pca(color_by = "class", frame = FALSE) +
  ggsci::scale_fill_lancet() +
  ggrepel::geom_text_repel(aes(label = ifelse(class == "QC", sample_id, NA)))
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
