% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massqc_pca.R
\name{massqc_pca_pc1}
\alias{massqc_pca_pc1}
\title{massqc_pca_pc1}
\usage{
massqc_pca_pc1(
  object,
  color_by,
  order_by,
  point_alpha = 0.8,
  point_size = 3,
  desc = FALSE
)
}
\arguments{
\item{object}{tidymass-class object.}

\item{color_by}{which column (sample_info) is used to color samples}

\item{order_by}{which column (sample_info) is used to order samples}

\item{point_alpha}{point_alpha}

\item{point_size}{point_size}

\item{desc}{desc}
}
\value{
ggplot2 plot.
}
\description{
Boxplot for each samples.
}
\examples{
library(massdataset)
library(ggplot2)
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )

object \%>\%
  massqc_pca_pc1()

object \%>\%
  massqc_pca_pc1(color_by = "class")

object \%>\%
  scale \%>\%
  massqc_pca_pc1(color_by = "class")

object \%>\%
  scale \%>\%
  massqc_pca_pc1(
    color_by = "class",
    order_by = "injection.order",
    point_alpha = 1,
    point_size = 5
  ) +
  ggsci::scale_color_lancet()

object \%>\%
  scale \%>\%
  massqc_pca_pc1(
    color_by = "class",
    order_by = "injection.order",
    point_alpha = 1,
    point_size = 5,
    desc = TRUE
  ) +
  ggsci::scale_color_lancet()
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
