% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massqc_report.R
\name{massqc_report}
\alias{massqc_report}
\title{Report data quality}
\usage{
massqc_report(object, path = ".", type = c("html", "pdf", "all"))
}
\arguments{
\item{object}{tidymass-class object.}

\item{path}{path}

\item{type}{type html, pdf or all.}
}
\value{
reporting result
}
\description{
Report data quality
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
