% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massqc_rsd_plot.R
\name{massqc_rsd_plot}
\alias{massqc_rsd_plot}
\title{massqc_rsd_plot}
\usage{
massqc_rsd_plot(
  object,
  color_by,
  order_by = "variable_id",
  show_x_text = FALSE,
  show_x_ticks = FALSE,
  desc = FALSE,
  point_alpha = 0.8,
  point_size = 3
)
}
\arguments{
\item{object}{tidymass-class object.}

\item{color_by}{which column (sample_info) is used to color samples}

\item{order_by}{which column (sample_info) is used to order samples}

\item{show_x_text}{show_x_text}

\item{show_x_ticks}{show_x_ticks}

\item{desc}{desc}

\item{point_alpha}{point_alpha}

\item{point_size}{point_size}
}
\value{
ggplot2 plot.
}
\description{
RSD plot for each variable.
}
\examples{
library(massdataset)
library(ggplot2)
library(tidyverse)
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )

object \%>\% 
  massqc_rsd_plot()

object \%>\% 
  massqc_rsd_plot(color_by = "rsd")

object \%>\% 
  massqc_rsd_plot(color_by = "rsd", order_by = "rsd")

object \%>\% 
  massqc_rsd_plot(color_by = "rsd", point_alpha = 1) +
  scale_color_gradient(low = "skyblue", high = "red") +
  geom_hline(yintercept = 0.3, color = "red")

object \%>\% 
  activate_mass_dataset(what = "sample_info") \%>\% 
  filter(class == "Subject") \%>\% 
  massqc_rsd_plot(color_by = "rsd", point_alpha = 1) +
  scale_color_gradient(low = "skyblue", high = "red") +
  geom_hline(yintercept = 0.3, color = "red")
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
