% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massqc_sample_correlation.R
\name{massqc_sample_correlation}
\alias{massqc_sample_correlation}
\title{massqc_sample_correlation}
\usage{
massqc_sample_correlation(
  object,
  cor_method = c("spearman", "kendall", "pearson"),
  method = c("circle", "square"),
  type = c("full", "lower", "upper"),
  ggtheme = ggplot2::theme_bw,
  title = "",
  show.legend = TRUE,
  legend.title = "Correlation",
  show.diag = FALSE,
  colors = c("blue", "white", "red"),
  outline.color = "gray",
  hc.order = FALSE,
  order_by,
  hc.method = "complete",
  lab = FALSE,
  lab_col = "black",
  lab_size = 4,
  p.mat = NULL,
  sig.level = 0.05,
  insig = c("pch", "blank"),
  pch = 4,
  pch.col = "black",
  pch.cex = 5,
  tl.cex = 12,
  tl.col = "black",
  tl.srt = 45,
  digits = 2
)
}
\arguments{
\item{object}{tidymass-class object.}

\item{cor_method}{?cor}

\item{method}{?ggcorrplot}

\item{type}{?ggcorrplot}

\item{ggtheme}{?ggcorrplot}

\item{title}{?ggcorrplot}

\item{show.legend}{?ggcorrplot}

\item{legend.title}{?ggcorrplot}

\item{show.diag}{?ggcorrplot}

\item{colors}{?ggcorrplot}

\item{outline.color}{?ggcorrplot}

\item{hc.order}{?ggcorrplot}

\item{order_by}{?ggcorrplot}

\item{hc.method}{?ggcorrplot}

\item{lab}{?ggcorrplot}

\item{lab_col}{?ggcorrplot}

\item{lab_size}{?ggcorrplot}

\item{p.mat}{?ggcorrplot}

\item{sig.level}{?ggcorrplot}

\item{insig}{?ggcorrplot}

\item{pch}{?ggcorrplot}

\item{pch.col}{?ggcorrplot}

\item{pch.cex}{?ggcorrplot}

\item{tl.cex}{?ggcorrplot}

\item{tl.col}{?ggcorrplot}

\item{tl.srt}{?ggcorrplot}

\item{digits}{?ggcorrplot}
}
\value{
ggplot2 plot.
}
\description{
Boxplot for each samples.
}
\examples{
library(massdataset)
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )

object \%>\%
  massqc_sample_correlation()

object \%>\%
  massqc_sample_correlation()

object \%>\%
  activate_mass_dataset(what = "sample_info") \%>\%
  filter(class == "QC") \%>\%
  massqc_sample_correlation()

object \%>\%
  massqc_sample_correlation(lab = TRUE)

object \%>\%
  massqc_sample_correlation(lab = TRUE, type = "upper")
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
