% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_boxplot.R
\name{sample_boxplot}
\alias{sample_boxplot}
\title{sample_boxplot}
\usage{
sample_boxplot(object, color_by, outlier.shape = NA)
}
\arguments{
\item{object}{tidymass-class object.}

\item{color_by}{one column of sample_info from object.}

\item{outlier.shape}{outlier.shape from ggplot2.}
}
\value{
ggplot2 plot.
}
\description{
Boxplot for each samples.
}
\examples{
library(massdataset)
library(ggplot2)
data("expression_data")
data("sample_info")
data("sample_info_note")
data("variable_info")
data("variable_info_note")
object =
create_tidymass_class(
expression_data = expression_data,
sample_info = sample_info,
variable_info = variable_info,
sample_info_note = sample_info_note,
variable_info_note = variable_info_note
)
object
object@expression_data = log(object@expression_data, 10)
plot =
sample_boxplot(object = object) +
geom_jitter(aes(color = class)) +
ggsci::scale_color_aaas()
plot
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
