% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{show_missing_values}
\alias{show_sample_missing_values}
\alias{show_variable_missing_values}
\alias{get_mv_number}
\alias{extract_sample_info}
\alias{extract_variable_info}
\alias{extract_expression_data}
\alias{extract_process_info}
\alias{show_mz_rt_plot}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{massdataset}{\code{\link[massdataset:mass_dataset-extracting]{extract_expression_data}}, \code{\link[massdataset:mass_dataset-extracting]{extract_process_info}}, \code{\link[massdataset:mass_dataset-extracting]{extract_sample_info}}, \code{\link[massdataset:mass_dataset-extracting]{extract_variable_info}}, \code{\link[massdataset:mass_dataset-summary]{get_mv_number}}, \code{\link[massdataset]{show_missing_values}}, \code{\link[massdataset]{show_mz_rt_plot}}, \code{\link[massdataset]{show_sample_missing_values}}, \code{\link[massdataset]{show_variable_missing_values}}}
}}

